[experiment/wip] module ecdsa_sign_to_contract, anti nonce sidechan
General sign to contract, with special util functions to perform the
anti nonce covert channel protection.

This is based on the description of the fix by Stepan: https://medium.com/cryptoadvance/hardware-wallets-can-be-hacked-but-this-is-fine-a6156bbd199

The protocol wording and some functions are copied/adapted from Jonas
Nick's PRs which do the same for BIP-Schnorr.

1.
Add test_ecdsa_s2c_anti_nonce_covert_channel_client_commit to return the
curve point committing to the signing client nonce.

This is a convenience function and can technically be emulated by
calling secp256k1_ecdsa_s2c_sign() and reconstructing the curve point from
the signature r/s values.